Added `roll dice` functionality.
User can use the command `!roll` to trigger the bot to roll dice.
The argument has to be in format
	`<num_dice>d<type_of_dice>[<modifier>]`

**Valid** examples are:
	* `!roll 5d20`
	* `!roll 2d5-5`
	* `!roll 4d6+2!`

The type of dice don't have to represent real-world dice, so a d5 is
a valid type. The modifier can be positive as well as negative.
The `!` in the last example signifies that the modifier should be
applied to each individual roll. If left out, the modifier will
be applied to the result of all rolls combined.

**Invalid** examples are:
	* `!roll 0d3`
	* `!roll 3d0`
	* `!roll 2d4 +4`
	* `!roll 4d2 + 3`

An amount of zero dice as well as dice of type 0 are invalid.
No whitespaces are allowed either between the type of dice and the
modifier or within the modifier itself.